Add more tests for conditions

+++ src/main/java/org/graphwalker/core/condition/EdgeCoverage.java
@@ -37,12 +37,12 @@ public final class EdgeCoverage implements StopCondition {
 
     private final double percent;
 
-    public EdgeCoverage(long percent) {
-        this.percent = percent/100;
+    public EdgeCoverage(int percent) {
+        this.percent = (double)percent/100;
     }
 
-    public double getPercent() {
-        return percent * 100;
+    public int getPercent() {
+        return (int)(percent * 100);
     }
 
     @Override
